v1.0.5 (#58)
* adding animate support to plot_pdb()

* typos

* bugfixes to and making block_diag default for homo-oligomeric design

* typo

* cleanup

* cleanup

* typo

* cleaning up output, moving self.log, self.loss, self.grad to self.aux[*]

* bugfix in trdesign

 was using best loss instead of current loss in callback function

* code cleanup

* adding predict() option

* typos

* moving model setup to prep, to allow changing configs during prep_inputs()

* typos

* typo

* adding documentation

* adding documentation

* adding experimental repeat support to partial

* cleanup: removing unused option from binder notebook

* adding experimental cropping from Shihao

* typos

* improvements

* adding experimental cmap

* adding rolling average to pAE plots when crops used

* bugfix: binder redesign thanks @blacktanktop

* add toxin design demo

* add WARNING if params not found

* updating path check

* renaming models->num_models, recycles->num_recycles to be consistent through the code

* cleaning up

* adding cropping based on distance

* Update and rename toxin_design.ipynb to desulfide_design.ipynb

* Update and rename desulfide_design.ipynb to disulfide_design.ipynb

* Update disulfide_design.ipynb

* Update disulfide_design.ipynb

* moving disulfide loss to custom loss

* fixing disulfide notebook

* cleaning up the warnings

* v1.0.5-beta - refactoring code to allow model._batch to change without recompiling (now part of model._inputs["batch"])

* v1.0.5-alpha

* adding option to set state of optimize within set_seq()

* predict() bugfix

* cleanup

* typo

* moving self._batch to self._inputs["batch"]

* moving atom37_to_frames inside compiled model (to avoid slow down during prep)

* bugfix for binder

* adding ptm

* fixing import

* minor edit

* Created using Colaboratory

* rm_template_sc: adding option to control if template sidechains are used as input

* bugfix

* typo

* Created using Colaboratory

* Created using Colaboratory

* Update README.md

* typo

* bugfix

* fixing link in AF2rank

+++ colabdesign/shared/utils.py
@@ -5,7 +5,8 @@ import jax.numpy as jnp
 
 def clear_mem():
   backend = jax.lib.xla_bridge.get_backend()
-  for buf in backend.live_buffers(): buf.delete()
+  if hasattr(backend,'live_buffers'):
+    for buf in backend.live_buffers(): buf.delete()
 
 def update_dict(D, *args, **kwargs):
   '''robust function for updating dictionary'''
@@ -43,10 +44,11 @@ def to_float(x):
   else: x = float(x)
   return x
 
-def dict_to_str(x, filt=None, keys=None, ok=None, print_str="", f=2):
+def dict_to_str(x, filt=None, keys=None, ok=None, print_str=None, f=2):
   '''convert dictionary to string for print out'''  
   if keys is None: keys = []
   if filt is None: filt = {}
+  if print_str is None: print_str = ""
   if ok is None: ok = ""
 
   # gather keys
